tray/requirements.txt
pystray>=0.19.0
psutil>=5.9.0
Pillow>=10.0.0
